+++ CHANGELOG.md
@@ -3,7 +3,6 @@
 ### v0.1.4
 
 - Add CdnClusterGet and CndNodeGet methods
+- Add Address and PrivateKeyHex method to crypto
 - Added Partition methods for ring topology
 - Updated bucket creating logic for mock DDC Smart-Contract
-
-